Add quick start tutorial with latest steps for creating supply chain graph
