+++ cmd/ratify/cmd/referrer.go
@@ -23,14 +23,11 @@ import (
 	"strings"
 
 	"github.com/deislabs/ratify/config"
-	"github.com/deislabs/ratify/pkg/common"
 	"github.com/deislabs/ratify/pkg/ocispecs"
-	"github.com/deislabs/ratify/pkg/referrerstore"
 	sf "github.com/deislabs/ratify/pkg/referrerstore/factory"
 	"github.com/deislabs/ratify/pkg/utils"
 	v1 "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/spf13/cobra"
-	"github.com/xlab/treeprint"
 )
 
 const (
@@ -56,38 +53,8 @@ func NewCmdReferrer(argv ...string) *cobra.Command {
 		},
 	}
 
-	cmd.AddCommand(NewCmdReferrerList(argv...))
 	cmd.AddCommand(NewCmdShowBlob(argv...))
 	cmd.AddCommand(NewCmdShowRefManifest(argv...))
-	return cmd
-}
-
-func NewCmdReferrerList(argv ...string) *cobra.Command {
-	var opts referrerCmdOptions
-
-	if len(argv) == 0 {
-		argv = []string{os.Args[0]}
-	}
-
-	eg := fmt.Sprintf(`  # List referrers for a subject
-  %s list -c ./config.yaml -s myregistry/myrepo@sha256:34343`, strings.Join(argv, " "))
-
-	cmd := &cobra.Command{
-		Use:     "list [OPTIONS]",
-		Short:   "List referrers to a subject",
-		Example: eg,
-		Args:    cobra.NoArgs,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			return listReferrers(opts)
-		},
-	}
-
-	flags := cmd.Flags()
-
-	flags.StringVarP(&opts.subject, "subject", "s", "", "Subject Reference")
-	flags.StringVarP(&opts.configFilePath, "config", "c", "", "Config File Path")
-	flags.StringArrayVarP(&opts.artifactTypes, "artifactType", "t", nil, "artifact type to filter")
-	flags.BoolVar(&opts.flatOutput, "flat", false, "Output referrers in a flat list format (default is tree format)")
 	return cmd
 }
 
@@ -147,104 +114,6 @@ func NewCmdShowRefManifest(argv ...string) *cobra.Command {
 	flags.StringVarP(&opts.digest, "digest", "d", "", "blob digest")
 	flags.StringVar(&opts.storeName, "store", "", "store name")
 	return cmd
-}
-
-type listResult struct {
-	Name       string                         `json:"storeName"`
-	References []ocispecs.ReferenceDescriptor `json:"References,omitempty"`
-}
-
-func Test(subject string) {
-	listReferrers((referrerCmdOptions{
-		subject:       subject,
-		artifactTypes: []string{"myartifact"},
-	}))
-}
-
-func listReferrers(opts referrerCmdOptions) error {
-
-	if opts.subject == "" {
-		return errors.New("subject parameter is required")
-	}
-
-	subRef, err := utils.ParseSubjectReference(opts.subject)
-	if err != nil {
-		return err
-	}
-
-	cf, err := config.Load(opts.configFilePath)
-	if err != nil {
-		return err
-	}
-
-	// TODO replace with code
-	rootImage := treeprint.NewWithRoot(subRef.String())
-
-	stores, err := sf.CreateStoresFromConfig(cf.StoresConfig, config.GetDefaultPluginPath())
-
-	if err != nil {
-		return err
-	}
-
-	type Result struct {
-		Name       string
-		References []ocispecs.ReferenceDescriptor
-	}
-
-	var results []listResult
-
-	for _, referrerStore := range stores {
-		storeNode := rootImage.AddBranch(referrerStore.Name())
-		result, err := listReferrersForStore(subRef, opts.artifactTypes, referrerStore, storeNode)
-		if err != nil {
-			return err
-		}
-		results = append(results, *result)
-	}
-
-	if !opts.flatOutput {
-		fmt.Println(rootImage.String())
-		return nil
-	}
-
-	return PrintJSON(results)
-}
-
-func listReferrersForStore(subRef common.Reference, artifactTypes []string, store referrerstore.ReferrerStore, treeNode treeprint.Tree) (*listResult, error) {
-	var continuationToken string
-	result := listResult{
-		Name: store.Name(),
-	}
-
-	for {
-		lr, err := store.ListReferrers(context.Background(), subRef, artifactTypes, continuationToken)
-		if err != nil {
-			return nil, err
-		}
-
-		continuationToken = lr.NextToken
-		for _, ref := range lr.Referrers {
-			refNode := treeNode.AddBranch(fmt.Sprintf("[%s]%s", ref.ArtifactType, ref.Digest.String()))
-			sr := common.Reference{
-				Path:     subRef.Path,
-				Digest:   ref.Digest,
-				Original: fmt.Sprintf("%s@%s", subRef.Path, ref.Digest),
-			}
-
-			subResult, err := listReferrersForStore(sr, artifactTypes, store, refNode)
-			if err != nil {
-				return nil, err
-			}
-			result.References = append(result.References, subResult.References...)
-
-		}
-		result.References = append(result.References, lr.Referrers...)
-		if continuationToken == "" {
-			break
-		}
-	}
-
-	return &result, nil
 }
 
 func showBlob(opts referrerCmdOptions) error {
